Updated to 1.13.1 and other changes Updated error/debug output Changed quote splitter to use regex Updated default values for metadata to better replicate a normal entity Fixed typo, removed unused calls Fixed arrows gaining particles TippedArrows will now show a random color instead of aqua AreaEffectCloud will now show a random color AreaEffectCloud's color methods now uses color not int Added org.bukkit.Color to the disguise parameters available as RGB or name Added new method ParticlesColor to LivingWatcher Fixed LivingWatcher's potion effect particles

+++ src/me/libraryaddict/disguise/disguisetypes/watchers/ArrowWatcher.java
@@ -10,7 +10,7 @@ public class ArrowWatcher extends FlagWatcher {
     }
 
     public boolean isCritical() {
-        return (byte) getData(MetaIndex.ARROW_CRITICAL) == 1;
+        return getData(MetaIndex.ARROW_CRITICAL) == 1;
     }
 
     public void setCritical(boolean critical) {
